# Remove respond_to? in assign_attribute happy path