--- a/activemodel/lib/active_model/attribute_assignment.rb
+++ b/activemodel/lib/active_model/attribute_assignment.rb
@@ -45,8 +45,10 @@ module ActiveModel
 
       def _assign_attribute(k, v)
         setter = :"#{k}="
+        public_send(setter, v)
+      rescue NoMethodError
         if respond_to?(setter)
-          public_send(setter, v)
+          raise
         else
           raise UnknownAttributeError.new(self, k.to_s)
         end